+++ src/Products/Ploneboard/content/PloneboardForum.py
@@ -82,10 +82,10 @@ schema = BaseBTreeFolderSchema + Schema((
         default=False,
         languageIndependent=0,
         widget = BooleanWidget(
-            label = u'Allow users to edit their comments',
-            description = u'If selected, this will give users the ability to edit their own comments.',
-            label_msgid = 'label_allow_edit_comment',
-            description_msgid = 'help_allow_edit_comment',
+            label = _('label_allow_edit_comment',
+                      default=u'Allow users to edit their comments'),
+            description = _('help_allow_edit_comment',
+                            default=u'If selected, this will give users the ability to edit their own comments.'),
             # Only show when no conversations exist
             condition = "not:object/getNumberOfConversations|nothing",
             ),